requirements.txt
discord.py==2.3.2
asyncpg==0.29.0
aiohttp==3.9.3
groq==0.9.0
python-dotenv==1.0.1
colorlog==6.8.2
cryptography==42.0.5
pydantic==2.7.1
tenacity==8.3.0
aiosqlite==0.20.0
